Refactor pool owner
